Fix: repository

+++ src/main/java/ru/gb/hw12/repository/TaskRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.gb.hw12.entity.Task;
 
 @Repository
-public interface TaskRepository extends JpaRepository<Task, Integer> {
+public interface TaskRepository extends JpaRepository<Task, Long> {
 }
